+++ tests/test_save_load.py
@@ -16,13 +16,12 @@ MODEL_LIST = [
     SAC,
 ]
 
-
 @pytest.mark.parametrize("model_class", MODEL_LIST)
 def test_save_load(model_class):
     """
     Test if 'save' and 'load' saves and loads model correctly
     and if 'load_parameters' and 'get_policy_parameters' work correctly
-    
+
     ''warning does not test function of optimizer parameter load
 
     :param model_class: (BaseRLModel) A RL model
@@ -38,16 +37,15 @@ def test_save_load(model_class):
     observations = np.squeeze(observations)
 
     # Get dictionary of current parameters
-    params = deepcopy(model.get_policy_parameters())
-    opt_params = deepcopy(model.get_opt_parameters())
+    params = deepcopy(model.policy.state_dict())
 
     # Modify all parameters to be random values
     random_params = dict((param_name, th.rand_like(param)) for param_name, param in params.items())
 
     # Update model parameters with the new random values
-    model.load_parameters(random_params, opt_params)
+    model.policy.load_state_dict(random_params)
 
-    new_params = model.get_policy_parameters()
+    new_params = model.policy.state_dict()
     # Check that all params are different now
     for k in params:
         assert not th.allclose(params[k], new_params[k]), "Parameters did not change as expected."
@@ -63,33 +61,14 @@ def test_save_load(model_class):
     model = model_class.load("test_save", env=env)
 
     # check if params are still the same after load
-    new_params = model.get_policy_parameters()
+    new_params = model.policy.state_dict()
 
     # Check that all params are the same as before save load procedure now
-    for k in params:
-        assert th.allclose(params[k], new_params[k]), "Model parameters not the same after save and load."
-
-    # check if optimizer params are still the same after load
-    new_opt_params = model.get_opt_parameters()
-    # check if keys are the same
-    assert opt_params.keys() == new_opt_params.keys()
-    # check if values are the same: only tested for Adam and RMSProp so far
-    # comparing states not implemented so far. hashes of state_entries are not the same for equal tensors
-    # comparing every sub_entry does not work because of bool value of Tensor with more than one value is ambiguous
-    # so far only comparing param_lists
-    for optimizer, opt_state in opt_params.items():
-        for param_group_idx, param_group in enumerate(opt_state['param_groups']):
-            for param_key, param_value in param_group.items():
-                # don't know how to handle params correctly, therefore only check if we have the same amount
-                if param_key == 'params':
-                    assert len(param_value) == len(
-                        new_opt_params[optimizer]['param_groups'][param_group_idx][param_key])
-                else:
-                    assert param_value == new_opt_params[optimizer]['param_groups'][param_group_idx][param_key]
+    for key in params:
+        assert th.allclose(params[key], new_params[key]), "Model parameters not the same after save and load."
 
     # check if model still selects the same actions
     new_selected_actions = [model.predict(observation, deterministic=True) for observation in observations]
-    # for i in range(len(selected_actionsselected_actions)):
     assert np.allclose(selected_actions, new_selected_actions)
 
     # check if learn still works