+++ docs/misc/changelog.rst
@@ -18,6 +18,7 @@ New Features:
 
 Bug Fixes:
 ^^^^^^^^^^
+- Fix loading model on CPU that were trained on GPU
 
 Deprecations:
 ^^^^^^^^^^^^^